Prototype (#38)
* Setting up the front end skeleton

* Add gulpfile and config, get skeleton app building. Remaining is to add layout HTML

* Add main UI layout to front end

* adding requirements.txt file

* update readme with back-end setup notes (#25)

* update readme with back-end setup notes

* fixing numbering

* Made UI changes for prototype

Not a finished product but it should be enough for the prototype.  A
major change in the UI in the future will be that picture previews will
appear above the choose file button.

* Setup Tutorial Controller/files

* Created Homebutton for tutorial page.

* Cleaned up buttons

* UI layout (#30)

* Made UI changes for prototype

Not a finished product but it should be enough for the prototype.  A
major change in the UI in the future will be that picture previews will
appear above the choose file button.

* Cleaned up buttons

* Forgot CSS

* Set up an image directory and work it into the GulpJS build process

* Adding opencv app skeleton (#36)

adding ccopencv app skeleton and url endpoint for front-end

* Tutorial images

* Making a couple changes in the README and the tutorial.

Just wanted to add some extra to make things more clear for the
prototype.